+++ app/page.js
@@ -1,13 +1,28 @@
+'use client'
 import Image from "next/image";
 import { Button } from '@mui/joy'
-import styles from "./page.module.css";
+import Alert from "@/components/alert";
+import styles from "@/styles/page.module.css";
 import Logo from "../public/images/FIS_CRESCENDO_Logo_Horizontal_RGB.png";
+import { useEffect } from "react";
 
 export default function Home() {
+  const isDevMode = process.env.NEXT_PUBLIC_DEV_MODE;
+  const isPostSeason = process.env.NEXT_PUBLIC_POSTSEASON;
+
+  useEffect(() => {
+    // if(isDevMode){ 
+      {isDevMode && isDevMode && <Alert mode='dev'/>}
+    // }
+    // if(isPostSeason){
+      {isPostSeason && isPostSeason && <Alert mode='postseason'/>}
+    // }
+  }, []);
+
   return (
     <>
     <header className={styles.flexHeader}>
-      <p>Version: <span id="version-number" style={{color: '#01a0bb'}}>2.1.1</span></p>
+      <p>Version: <span id="version-number" style={{color: '#01a0bb'}}>2024-4.0.0</span></p>
       <nav style={{display:'flex', flexDirection: 'column'}}>
         <Button component="a" href="/pit-results" sx={{mb:'0.5rem'}}>
           View Pit Results &#8594;